% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iface-class.R
\name{as.list.iface}
\alias{as.list.iface}
\title{Cast an \code{iface} to a plain list.}
\usage{
\method{as.list}{iface}(x, ..., flatten = FALSE)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{objects, possibly named.}

\item{flatten}{get a list of lists representation instead
of the dataframe column by column list.}
}
\value{
a list representation of the \code{iface} input.
}
\description{
Cast an \code{iface} to a plain list.
}
\examples{
my_iface = iface( 
  col1 = integer + group_unique ~ "an integer column"
)

as.list(my_iface, flatten=TRUE)
}
