% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_check_types.R
\name{check_date}
\alias{check_date}
\title{Checks a set of variables can be coerced to a date and coerces them}
\usage{
check_date(
  ...,
  .message = "`{param}` is not a date: ({err}).",
  .env = rlang::caller_env()
)
}
\arguments{
\item{...}{a list of symbols
  Named arguments passed on to \code{\link[base:as.Date]{base::as.Date}}\describe{
    \item{\code{x}}{an object to be converted.}
\item{\code{format}}{a \code{\link[base]{character}} string.  If not specified when
    converting from a character representation, it will try
    \code{tryFormats} one by one on the first non-\code{NA} element, and
    give an error if none works.  Otherwise, the processing is via
    \code{\link[base]{strptime}()} whose help page describes available
    conversion specifications.}
\item{\code{tryFormats}}{\code{\link[base]{character}} vector of \code{format}
    strings to try if \code{format} is not specified.}
\item{\code{optional}}{\code{\link[base]{logical}} indicating to return \code{NA}
    (instead of signalling an error) if the format guessing does not succeed.}
\item{\code{origin}}{a \code{Date} object, or something which can be coerced by
    \code{as.Date(origin, \dots)} to such an object or
    \code{\link[base]{missing}}.  In that case, \code{"1970-01-01"} is used.}
\item{\code{tz}}{a time zone name.}
}}

\item{.message}{a glue specification containing \code{{param}} as the name of the
parameter and \code{{err}} the cause of the error}

\item{.env}{the environment to check (defaults to calling environment)}
}
\value{
nothing. called for side effects. throws error if not all variables can be coerced.
}
\description{
Checks a set of variables can be coerced to a date and coerces them
}
\examples{
a = c(Sys.Date()+1:10)
b = format(a)
f = "1970-01-01"
g = NA
check_date(a,b,f,g)

c = c("dfsfs")
try(check_date(c,d, mean))
}
\concept{parameter_checks}
