% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-demo-functions.R
\name{demo_iparam_tag}
\alias{demo_iparam_tag}
\title{An example function}
\usage{
demo_iparam_tag(mydata = i_test, another = "value", ...)
}
\arguments{
\item{mydata}{description of a dataframe - format automatically documented - a dataframe with columns:
\itemize{
\item id (integer) - an integer ID
\item test (logical) - the test result
}

Any grouping allowed.
}

\item{another}{an example}

\item{...}{not used}
}
\value{
the conformant dataframe
}
\description{
An example function
}
\keyword{internal}
