% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-demo-functions.R
\name{demo_ireturn_tag}
\alias{demo_ireturn_tag}
\title{Another example function}
\usage{
demo_ireturn_tag(mydata = i_test_extn, ...)
}
\arguments{
\item{mydata}{a more constrained input - a dataframe with columns:
\itemize{
\item id (integer) - an integer ID
\item test (logical) - the test result
\item extra (character) - a new value
}

Ungrouped.
}

\item{...}{not used}

\item{another}{an example}
}
\value{
a test output, format will be automatically documented

A dataframe containing the following columns: 
\itemize{
\item id (integer) - an integer ID
\item test (logical) - the test result
}

Any grouping possible.
}
\description{
Another example function
}
\keyword{internal}
