% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{iclip}
\alias{iclip}
\title{Create an \code{iface} interface from an example dataframe}
\usage{
iclip(df, df_name = deparse(substitute(df)))
}
\arguments{
\item{df}{a prototype dataframe}

\item{df_name}{an optional name for the parameter}
}
\value{
nothing
}
\description{
Copies the \code{iface} specification to the clipboard
}
\examples{
if(FALSE) iclip(iris)
}
\concept{document}
