% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfacer.R
\name{iface}
\alias{iface}
\title{Define an interface}
\usage{
iface(...)
}
\arguments{
\item{...}{The specification of the interface (see details)}
}
\value{
the definition of an interface as a \code{iface} object
}
\description{
This is the default value for a dataframe.
The specificiation is in the form of a named list of formulae with
the structure \code{name = type ~ "documentation"}
}
\examples{
my_iface = iface(col1 = integer ~ "an integer column" )
print(my_iface)

x = function(df = my_iface, ...) {
  df = ivalidate(df,...)
  return(df)
}

x(tibble::tibble(col1 = c(1,2,3)))

my_iface2 = iface(my_iface, col2 = character ~ "another col")
print(my_iface2)
}
