% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfacer.R
\name{imapper}
\alias{imapper}
\title{Specify mappings that can make dataframes compatible with an interface}
\usage{
imapper(...)
}
\arguments{
\item{...}{a set of \code{dplyr::mutate()} specifications that when applied to
a dataframe will rename or otherwise fix missing columns}
}
\value{
a set of mappings
}
\description{
This function is expected to be used only in a \code{.imap = imappper(...)} context
to overcome some mapping issues
}
\examples{
x = function(df = iface(col1 = integer ~ "an integer column" ), ...) {
  df = ivalidate(df,...)
}
input=tibble::tibble(col2 = c(1,2,3)) 
# This fails because col1 is missing
try(x(input))
# This fixes it for this input
x(input, .imap=imapper(col1 = col2))
}
\concept{interface}
