% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfacer.R
\name{in_range}
\alias{in_range}
\title{Define a conformance rule to confirm that a numeric is in a set range}
\usage{
in_range(min, max)
}
\arguments{
\item{min}{the lower limit (inclusive)}

\item{max}{the upper limit (inclusive)}
}
\value{
a function which checks the values and returns them if OK or throws
an error if not
}
\description{
This is anticipated to be part of a \code{iface} rule e.g.
}
\details{
\code{iface(test_col = integer + in_range(-10,10) ~ "An integer from -10 to 10")}
}
\examples{
in_range(0,100)(1:99)
try(in_range(0,10)(1:99))
}
\concept{rules}
