% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfacer-package.R
\docType{package}
\name{interfacer-package}
\alias{interfacer}
\alias{interfacer-package}
\title{interfacer: Define and Enforce Contracts for Dataframes as Function Parameters}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A dataframe validation framework for package builders who use dataframes as function parameters. It performs checks on column names, coerces data-types, and checks grouping to make sure user inputs conform to a specification provided by the package author. It provides a mechanism for package authors to automatically document supported dataframe inputs and selectively dispatch to functions depending on the format of a dataframe much like S3 does for classes. It also contains some developer tools to make working with and documenting dataframe specifications easier. It helps package developers to improve their documentation and simplifies parameter validation where dataframes are used as function parameters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ai4ci.github.io/interfacer/}
  \item \url{https://github.com/ai4ci/interfacer}
  \item Report bugs at \url{https://github.com/ai4ci/interfacer/issues}
}

}
\author{
\strong{Maintainer}: Robert Challen \email{rob.challen@bristol.ac.uk} (\href{https://orcid.org/0000-0002-5504-7768}{ORCID}) [copyright holder]

}
\keyword{internal}
