% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iproto.R
\name{iproto}
\alias{iproto}
\title{Generate a zero length tibble conforming to a spec}
\usage{
iproto(iface)
}
\arguments{
\item{iface}{the specification}
}
\value{
a tibble conforming to \code{iface}
}
\description{
Generate a zero length tibble conforming to a spec
}
\examples{
i = interfacer::iface(
  col1 = integer ~ "A number",
  col2 = character ~ "A string"
)

iproto(i)
}
\concept{interface}
