% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfacer.R
\name{is.iface}
\alias{is.iface}
\title{Check if an object is an interface spec}
\usage{
is.iface(x, ...)
}
\arguments{
\item{x}{the parameter to check}

\item{...}{ignored}
}
\value{
a boolean.
}
\description{
Check if an object is an interface spec
}
\concept{interface}
