% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfacer.R
\name{itest}
\alias{itest}
\title{Test dataframe conformance to an interface specification}
\usage{
itest(df, iface = NULL, .imap = imap())
}
\arguments{
\item{df}{a dataframe to test}

\item{iface}{an interface specification produced by \code{iface()}. If missing
this will be inferred from the current function signature.}

\item{.imap}{an optional mapping specification produced by \code{imap()}}
}
\value{
TRUE if the dataframe is conformant, FALSE otherwise
}
\description{
\code{ivalidate} throws errors deliberately however sometimes dealing with invalid
input may be possible.
}
\examples{
i_diamonds = iface( 
  color = enum(D,E,F,G,H,I,J,extra) ~ "the colour", 
  price = integer ~ "the price"
)
itest(ggplot2::diamonds, i_diamonds)
}
