% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iface-class.R
\name{knit_print.iface}
\alias{knit_print.iface}
\title{Format an \code{iface} specification for printing}
\usage{
knit_print.iface(x, ...)
}
\arguments{
\item{x}{an \code{iface} specification}

\item{...}{not used.}
}
\value{
a formatted string representation of an \code{iface}
}
\description{
Format an \code{iface} specification for printing
}
\examples{
my_iface = iface( 
  col1 = integer + group_unique ~ "an integer column"
)

print(my_iface)
knitr::knit_print(my_iface)
}
