% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{roxy_tag_parse.roxy_tag_iparam}
\alias{roxy_tag_parse.roxy_tag_iparam}
\title{Parser for \verb{@iparam} tags}
\usage{
\method{roxy_tag_parse}{roxy_tag_iparam}(x)
}
\arguments{
\item{x}{A tag}
}
\description{
Parser for \verb{@iparam} tags
}
\concept{document}
