% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{roxy_tag_parse.roxy_tag_ireturn}
\alias{roxy_tag_parse.roxy_tag_ireturn}
\title{Parser for \verb{@ireturn} tags}
\usage{
\method{roxy_tag_parse}{roxy_tag_ireturn}(x)
}
\arguments{
\item{x}{A tag}
}
\value{
a \code{roxy_tag} object with the \code{val} field set to the parsed value
}
\description{
The \verb{@ireturn <description>} tag can be used in \code{roxygen2} documentation
of a function to describe a dataframe return value. The function must be using
\code{interfacer::ireturn} to define the output dataframe format. The
\verb{@ireturn} tag will then generate documentation about the type of dataframe
the function outputs.
}
\examples{
# This provides support to `roxygen2` and only gets executed in the context
# of `devtools::document()`. There is no interactive use of this function.
}
\concept{document}
