% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.anything}
\alias{type.anything}
\title{Coerce to an unspecified type}
\usage{
type.anything(x)
}
\arguments{
\item{x}{any vector}
}
\value{
the input (unless x is \code{NULL} in which case a \code{character()})
}
\description{
Coerce to an unspecified type
}
\concept{rules}
