% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.complete}
\alias{type.complete}
\title{Coerce to a complete set of values.}
\usage{
type.complete(x)
}
\arguments{
\item{x}{any vector, factor or numeric}
}
\value{
the input or error if not complete
}
\description{
This test checks either for factors that all factor levels are present in the
input, or for numerics if the sequence from minimum to maximum by the
smallest difference are not all (approximately) present. Empty values are
ignored.
}
\concept{rules}
