% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.date}
\alias{type.date}
\title{Coerce to a Date.}
\usage{
type.date(x, ...)
}
\arguments{
\item{x}{an object to be converted.}

\item{...}{further arguments to be passed from or to other methods,
    including \code{format} for \code{as.character} and \code{as.Date}
    methods.}
}
\description{
Coerce to a Date.
}
\concept{rules}
