% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.default}
\alias{type.default}
\title{Set a default value for a column}
\usage{
type.default(value)
}
\arguments{
\item{value}{a length one item of the correct type.}
}
\value{
a validation function that switches NAs for default values
}
\description{
Any NA values will be replaced by this value. N.b. default values must be
provided before any other rules if the validation is not to fail.
}
\concept{rules}
