% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.double}
\alias{type.double}
\title{Coerce to a double.}
\usage{
type.double()
}
\value{
the input as a double, error if this would involve data loss.
}
\description{
Coerce to a double.
}
\concept{rules}
