% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.factor}
\alias{type.factor}
\title{Coerce to a factor.}
\usage{
type.factor(x)
}
\arguments{
\item{x}{any vector}
}
\value{
the input as a factor, error if this would involve data loss.
}
\description{
Coerce to a factor.
}
\concept{rules}
