% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.finite}
\alias{type.finite}
\title{Check for non-finite values}
\usage{
type.finite(x)
}
\arguments{
\item{x}{any vector that can be coerced to numeric}
}
\value{
the numeric value of \code{x} if all finite values detected
}
\description{
Any non finite values will cause failure of validation.
}
\concept{rules}
