% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.group_unique}
\alias{type.group_unique}
\title{Coerce to a unique value within the current grouping structure.}
\usage{
type.group_unique(x)
}
\arguments{
\item{x}{any vector}
}
\value{
the input or error if any of x is not unique.
}
\description{
Coerce to a unique value within the current grouping structure.
}
\concept{rules}
