% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.integer}
\alias{type.integer}
\title{Coerce to integer}
\usage{
type.integer(x)
}
\arguments{
\item{x}{any vector}
}
\value{
the input as an integer, error if this would involve data loss.
}
\description{
Coerce to integer
}
\concept{rules}
