% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.logical}
\alias{type.logical}
\title{Coerce to a logical}
\usage{
type.logical(x)
}
\arguments{
\item{x}{any vector}
}
\value{
the input as a logical, error if this would involve data loss.
}
\description{
Coerce to a logical
}
\concept{rules}
