% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.not_missing}
\alias{type.not_missing}
\title{Check for missing values}
\usage{
type.not_missing(x)
}
\arguments{
\item{x}{any vector, factor or numeric}
}
\value{
\code{x} if no missing values detected
}
\description{
Any NA values will cause failure of validation.
}
\concept{rules}
