% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.numeric}
\alias{type.numeric}
\title{Coerce to a numeric.}
\usage{
type.numeric(x)
}
\arguments{
\item{x}{any vector}
}
\value{
the input as a numeric, error if this would involve data loss.
}
\description{
Coerce to a numeric.
}
\concept{rules}
