% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.of_type}
\alias{type.of_type}
\title{Check for a given class}
\usage{
type.of_type(type, .not_null = FALSE)
}
\arguments{
\item{type}{the class of the type we are checking as a symbol}

\item{.not_null}{are NULL values allowed (for list column entries only)}
}
\value{
a function that can check the input is of the correct type.
}
\description{
Any values of the wrong class will cause failure of validation. This is
particularly useful for custom vectors of for list types (e.g. \code{list(of_type(lm))})
}
\concept{rules}
