% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.positive_double}
\alias{type.positive_double}
\title{Coerce to a positive double.}
\usage{
type.positive_double(x)
}
\arguments{
\item{x}{object to be coerced or tested.}
}
\value{
the input as a positive double, error if this would involve data loss.
}
\description{
Coerce to a positive double.
}
\concept{rules}
