% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.positive_integer}
\alias{type.positive_integer}
\title{Coerce to a positive integer.}
\usage{
type.positive_integer(x)
}
\arguments{
\item{x}{any vector}
}
\value{
the input as a positive integer, error if this would involve data loss.
}
\description{
Coerce to a positive integer.
}
\concept{rules}
