% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.proportion}
\alias{type.proportion}
\title{Coerce to a number between 0 and 1}
\usage{
type.proportion(x)
}
\arguments{
\item{x}{object to be coerced or tested.}
}
\value{
the input as a number from 0 to 1, error if this would involve data loss.
}
\description{
Coerce to a number between 0 and 1
}
\concept{rules}
