% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.unique_id}
\alias{type.unique_id}
\title{A globally unique ids.}
\usage{
type.unique_id(x)
}
\arguments{
\item{x}{any vector}
}
\value{
the input.
}
\description{
A globally unique ids.
}
\concept{rules}
