% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{use_iface}
\alias{use_iface}
\title{Generate interfacer code for a dataframe}
\usage{
use_iface(
  df,
  name = deparse(substitute(df)),
  output = "R/interfaces.R",
  use_as_default = FALSE,
  pkg = "."
)
}
\arguments{
\item{df}{the data frame to use}

\item{name}{the name of the variable you wish to use (defaults to whatever
the function is called with)}

\item{output}{where to write data documentation code (defaults to \code{R/data.R})}

\item{use_as_default}{if this is set to true the current dataframe is saved
as package data and the \code{interfacer::iface} specification is created
referring to the package copy of the current dataframe as the default
value.}

\item{pkg}{the package (defaults to current)}
}
\value{
nothing
}
\description{
Generating and documenting an \code{iface} for a given dataframe would be time
consuming and annoying if you can't do it automatically. In this case as you
interactively develop a package any given dataframe's structure can be
explicitly documented and made into a specific contract within the package.
}
\concept{document}
