% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{iclip}
\alias{iclip}
\title{Create an \code{iface} interface from an example dataframe}
\usage{
iclip(df)
}
\arguments{
\item{df}{a prototype dataframe}
}
\value{
nothing
}
\description{
Copies to the clipboard
}
\examples{
if(FALSE) iclip(iris)
}
