% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{idocument}
\alias{idocument}
\title{Document an interface contract for inserting into \code{roxygen2}}
\usage{
idocument(fn, param = NULL)
}
\arguments{
\item{fn}{the function that you are documenting}

\item{param}{the parameter you are documenting (optional. if missing defaults
to the first argument of the function)}
}
\value{
a markdown snippet
}
\description{
This function is expected to be called within the documentation of a function
as inline code in the parameter documentation of the function. It details the
expected columns that the input dataframe should possess. This has mostly
been superseded by the \verb{@iparam <name> <description>} \code{roxygen2} tag which does
this automatically, however in some circumstances (particularly multiple
dispatch) you may want to assemble dataframe documentation manually.
}
\examples{
#' @param df `r idocument(x, df)`
x = function(df = iface( col1 = integer ~ "an integer column" )) {}

cat(idocument(x, df))
}
\concept{document}
