% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfacer.R
\name{iface}
\alias{iface}
\title{Define an interface}
\usage{
iface(...)
}
\arguments{
\item{...}{The specification of the interface (see details), or an unnamed
\code{iface} object to extend, or both.}
}
\value{
the definition of an interface as a \code{iface} object
}
\description{
This defines the structure of a dataframe.
The specification is in the form of a named list of formulae with
the structure \code{name = type ~ "documentation"}. \code{type} can be one of
'integer','double','numeric','date','logical','factor','character','primary_key' or \code{enum(level1,level2,...)} or
anything that resolves to a function e.g. \code{as.ordered}
}
\examples{
my_iface = iface(col1 = integer ~ "an integer column" )
print(my_iface)

x = function(df = my_iface, ...) {
  df = ivalidate(df,...)
  return(df)
}

x(tibble::tibble(col1 = c(1,2,3)))

my_iface2 = iface(my_iface, col2 = character ~ "another col")
print(my_iface2)
}
