% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iface-class.R
\name{is.iface}
\alias{is.iface}
\title{Check if an object is an interface specification}
\usage{
is.iface(x, ...)
}
\arguments{
\item{x}{the object to check}

\item{...}{ignored}
}
\value{
a boolean.
}
\description{
Check if an object is an interface specification
}
\concept{interface}
