% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfacer.R
\name{ivalidate}
\alias{ivalidate}
\title{Perform interface checks on dataframe by looking at enclosing function}
\usage{
ivalidate(df, ..., .imap = imap(), prune = FALSE)
}
\arguments{
\item{df}{a dataframe}

\item{...}{not used}

\item{.imap}{a set of mappings as an \code{imap} object}

\item{prune}{get rid of excess columns that are not in the spec}
}
\value{
a dataframe based on df with
}
\description{
Perform interface checks on dataframe by looking at enclosing function
}
\examples{
x = function(df = iface(col1 = integer ~ "an integer column" ), ...) {
  df = ivalidate(df,...)
  return(df)
}
input=tibble::tibble(col1 = c(1,2,3)) 
x(input)

# This fails because col1 is not coercable to integer
input2=tibble::tibble(col1 = c(1.5,2,3)) 
try(x(input2))
}
