% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{roxy_tag_parse.roxy_tag_iparam}
\alias{roxy_tag_parse.roxy_tag_iparam}
\title{Parser for \verb{@iparam} tags}
\usage{
\method{roxy_tag_parse}{roxy_tag_iparam}(x)
}
\arguments{
\item{x}{A tag}
}
\value{
a \code{roxy_tag} object with the \code{val} field set to the parsed value
}
\description{
The \verb{@iparam <name> <description>} tag can be used in \code{roxygen2} documentation
of a function to describe a dataframe parameter. The function must be using
\code{interfacer::iface} to define the input dataframe parameter format. The
\verb{@iparam} tag will then generate documentation about the type of dataframe
the function is expecting.
}
\examples{
# This provides support to `roxygen2` and only gets executed in the context
# of `devtools::document()`. There is no interactive use of this function.
}
\concept{document}
