% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-coerce.R
\name{type.date}
\alias{type.date}
\title{Coerce to a Date.}
\usage{
type.date(x, ...)
}
\arguments{
\item{x}{an object to be converted.}

\item{...}{further arguments to be passed from or to other methods.}
}
\value{
the input as a \code{date} vector, error if this would involve data loss.
}
\description{
Coerce to a Date.
}
\concept{rules}
